fix(backend): make int/uint genuine target-native 64-bit word types
int was hardcoded to int32_t despite the checker's LiteralTarget
already validating it as a 64-bit-range word type (WordBits: 64,
hardcoded in cmd/pebc and cmd/tirdump — this compiler only ever
targets a 64-bit host today). spec/compiler/05-types-and-inference.md
is explicit: int/uint are the target's native word types, 32-bit on a
32-bit target and 64-bit on a 64-bit target. cType now maps
types.Int to int64_t (uint's int64_t/uint64_t mapping was already
correct by coincidence). Threaded consistently through
printfSpecifier, integerKindRange, and the checked-arithmetic/shift/
negation/optional-unwrap/float-cast runtime-helper-suffix selection
functions, which previously routed int through the i32 helper family
purely because int and i32 happened to share a C representation.

A matching checker-side fix (isPointerWidthInteger, cTypeWidth) keeps
pointer-to-int-cast validation and structural width comparisons
consistent with the new backend mapping — int is now genuinely
pointer-width, and int/i64 (not int/i32) are the same concrete width.

Also fixes a related, previously-latent correctness gap found while
verifying: 64-bit signed integer literals need an explicit C `LL`
suffix (mirroring the existing `u` suffix for unsigned literals) to
avoid ambiguous/truncating literal typing, applying to both int and
i64. Widens one backend validation path that explicitly enumerated
admitted parameter kinds by name and had silently relied on int
matching i32's admission rather than being explicitly listed.

Verified end-to-end against the confirmed bug repro
(`let x int = 2147483648;`, previously failed cc with
-Wconstant-conversion, now compiles and runs correctly) plus a
causation check against HEAD. Full internal/backend and internal/check
suites clean.

+++ compiler/internal/backend/collect_test.go
@@ -29,15 +29,16 @@ func TestEmitGenericReachabilityEmitsThreeSpecializations(t *testing.T) {
 // TestEmitGenericHelperSpecializedAtConcreteWidthCompilesAndRuns is the exact
 // motivating repro for the compatible-integer-width parameter gate: a generic
 // helper identity[T] called with an i32 local from an fn main() int entry
-// (whose own resolved width is the abstract int builtin, NOT i32 — the two are
-// distinct builtins sharing the int32_t C representation). The specialization
+// (whose own resolved width is the 64-bit target-native int builtin, distinct
+// from i32 — the two share no C representation anymore). The specialization
 // identity[i32] has an i32-typed parameter, which the pre-fix
 // validateHelperSignature rejected with "called function symbol ... has type
 // i32, want int". The program must compile under -Wall -Wextra -Werror and run,
-// returning the value identity passed through.
+// returning the value identity passed through (widened back to the int return
+// with an explicit cast, since i32 no longer converts implicitly to int).
 func TestEmitGenericHelperSpecializedAtConcreteWidthCompilesAndRuns(t *testing.T) {
 	t.Parallel()
-	emitAndRun(t, `fn identity[T](x T) T { return x; } fn main() int { var a i32 = 5; var r = identity(a); return r; }`, false, 5, false)
+	emitAndRun(t, `fn identity[T](x T) T { return x; } fn main() int { var a i32 = 5; var r = identity(a); return r as int; }`, false, 5, false)
 }
 
 // TestEmitGenericClampShapeSpecializedAtConcreteWidthCompilesAndRuns mirrors
@@ -47,10 +48,11 @@ func TestEmitGenericHelperSpecializedAtConcreteWidthCompilesAndRuns(t *testing.T
 // concrete i32 width, called from an fn main() int entry with i32 locals and
 // the result stored into an i32 local before returning (the shape that reaches
 // emission — a direct `return clamp(...)` from an int entry hits a checker-level
-// int-vs-i32 unification conflict instead). The clamp of (5, 10, 20) is 10.
+// int-vs-i32 unification conflict instead). The clamp of (5, 10, 20) is 10,
+// widened back to the int return with an explicit cast.
 func TestEmitGenericClampShapeSpecializedAtConcreteWidthCompilesAndRuns(t *testing.T) {
 	t.Parallel()
-	emitAndRun(t, `fn min[T](a T, b T) T { if a < b { return a; } return b; } fn max[T](a T, b T) T { if a > b { return a; } return b; } fn clamp[T](x T, lo T, hi T) T { return max(lo, min(x, hi)); } fn main() int { var x i32 = 5; var lo i32 = 10; var hi i32 = 20; var r i32 = clamp(x, lo, hi); return r; }`, false, 10, false)
+	emitAndRun(t, `fn min[T](a T, b T) T { if a < b { return a; } return b; } fn max[T](a T, b T) T { if a > b { return a; } return b; } fn clamp[T](x T, lo T, hi T) T { return max(lo, min(x, hi)); } fn main() int { var x i32 = 5; var lo i32 = 10; var hi i32 = 20; var r i32 = clamp(x, lo, hi); return r as int; }`, false, 10, false)
 }
 
 func TestEmitRecursionWritesPrototypesBeforeDefinitions(t *testing.T) {
